Modificación de clases Persona

+++ src/main/java/com/portfolio/mm/Entity/Persona.java
@@ -11,7 +11,7 @@ import javax.validation.constraints.Size;
 public class Persona {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     
     @NotNull
     @Size(min = 1, max = 50, message = "no cumple con la longitud")
@@ -39,11 +39,11 @@ public class Persona {
     
     //Getter & Setter
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -78,6 +78,7 @@ public class Persona {
     public void setImg(String img) {
         this.img = img;
     }
+
     
 }
